+++ simpleTransformation.h
@@ -1,6 +1,6 @@
 #pragma once
 
-void add_simpleTransformation(matrix *A,int to,int what, int scalar, char orient );
-void mul_simpleTransformation(matrix *A,int scalar);
-void div_simpleTransformation(matrix *A,int scalar);
+void add_simpleTransformation(matrix *A,int to,int what, double scalar, char orient );
+void mul_simpleTransformation(matrix *A,double scalar);
+void div_simpleTransformation(matrix *A, double scalar);
 void swap_simpleTransformation(matrix *A,int what,int with,char orient );
